+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project (attempts to) adhere to [Semantic Versioning](http://semver.org/).
 
+## [1.2.11P] - Unpushed
+- Changes to speed up multiprocessing
+- gene2sequence is now a global
+
 ## [1.2.10] - 2020-04-03
 - Change the way that profile_genes works on the backend to optimize speed
 - Turn off complete scaffold-level read profiling by default; make it a command line option